+++ kripodb/canned.py
@@ -18,11 +18,27 @@ For using Kripo data files inside KNIME (http://www.knime.org)
 
 from __future__ import absolute_import
 
+import numpy as np
 import pandas as pd
+from requests import HTTPError
 
 from .db import FragmentsDb
 from .pairs import similar, open_similarity_matrix
-from .webservice.client import WebserviceClient
+from .webservice.client import WebserviceClient, IncompleteFragments
+
+
+class IncompleteHits(Exception):
+    def __init__(self, absent_identifiers, hits):
+        """List of hits and list of identifiers for which no information could be found
+
+        Args:
+            absent_identifiers (List[str]): List of identifiers for which no information could be found
+            hits (pandas.DataFrame): Data frame with query_fragment_id, hit_frag_id and score columns
+        """
+        message = 'Some query fragment identifiers could not be found'
+        super(IncompleteHits, self).__init__(message)
+        self.absent_identifiers = absent_identifiers
+        self.hits = hits
 
 
 def similarities(queries, similarity_matrix_filename_or_url, cutoff, limit=1000):
@@ -54,26 +70,48 @@ def similarities(queries, similarity_matrix_filename_or_url, cutoff, limit=1000)
 
     Returns:
         pandas.DataFrame: Data frame with query_fragment_id, hit_frag_id and score columns
+
+    Raises:
+        IncompleteHits: When one or more of the identifiers could not be found.
     """
     hits = []
+    absent_identifiers = []
     if similarity_matrix_filename_or_url.startswith('http'):
         client = WebserviceClient(similarity_matrix_filename_or_url)
         for query in queries:
-            qhits = client.similar_fragments(query, cutoff, limit)
-            hits.extend(qhits)
+            try:
+                qhits = client.similar_fragments(query, cutoff, limit)
+                hits.extend(qhits)
+            except HTTPError as e:
+                if e.response.status_code == 404:
+                    absent_identifiers.append(query)
     else:
         similarity_matrix = open_similarity_matrix(similarity_matrix_filename_or_url)
         for query in queries:
-            for query_id, hit_id, score in similar(query, similarity_matrix, cutoff, limit):
-                hit = {'query_frag_id': query_id,
-                       'hit_frag_id': hit_id,
-                       'score': score,
-                       }
-                hits.append(hit)
+            try:
+                for query_id, hit_id, score in similar(query, similarity_matrix, cutoff, limit):
+                    hit = {'query_frag_id': query_id,
+                           'hit_frag_id': hit_id,
+                           'score': score,
+                           }
+                    hits.append(hit)
+            except KeyError:
+                absent_identifiers.append(query)
 
         similarity_matrix.close()
 
-    return pd.DataFrame(hits)
+    if absent_identifiers:
+        if len(hits) > 0:
+            df = pd.DataFrame(hits, columns=['query_frag_id', 'hit_frag_id', 'score'])
+        else:
+            # empty hits array will give dataframe without columns
+            df = pd.DataFrame({'query_frag_id': pd.Series(dtype=str),
+                               'hit_frag_id': pd.Series(dtype=str),
+                               'score': pd.Series(dtype=np.double)
+                               }, columns=['query_frag_id', 'hit_frag_id', 'score'])
+        raise IncompleteHits(absent_identifiers, df)
+
+    return pd.DataFrame(hits, columns=['query_frag_id', 'hit_frag_id', 'score'])
 
 
 def fragments_by_pdb_codes(pdb_codes, fragments_db_filename_or_url, prefix=''):
@@ -104,16 +142,32 @@ def fragments_by_pdb_codes(pdb_codes, fragments_db_filename_or_url, prefix=''):
 
     Returns:
         pandas.DataFrame: Data frame with fragment information
+
+    Raises:
+        IncompleteFragments: When one or more of the identifiers could not be found.
     """
     if fragments_db_filename_or_url.startswith('http'):
         client = WebserviceClient(fragments_db_filename_or_url)
-        fragments = client.fragments_by_pdb_codes(pdb_codes)
+        try:
+            fragments = client.fragments_by_pdb_codes(pdb_codes)
+        except IncompleteFragments as e:
+            df = pd.DataFrame(e.fragments)
+            df.rename(columns=lambda x: prefix + x, inplace=True)
+            raise IncompleteFragments(e.absent_identifiers, df)
     else:
         fragmentsdb = FragmentsDb(fragments_db_filename_or_url)
         fragments = []
+        absent_identifiers = []
         for pdb_code in pdb_codes:
-            for fragment in fragmentsdb.by_pdb_code(pdb_code):
-                fragments.append(fragment)
+            try:
+                for fragment in fragmentsdb.by_pdb_code(pdb_code):
+                    fragments.append(fragment)
+            except LookupError as e:
+                absent_identifiers.append(pdb_code)
+        if absent_identifiers:
+            df = pd.DataFrame(fragments)
+            df.rename(columns=lambda x: prefix + x, inplace=True)
+            raise IncompleteFragments(absent_identifiers, df)
 
     df = pd.DataFrame(fragments)
     df.rename(columns=lambda x: prefix + x, inplace=True)
@@ -146,13 +200,31 @@ def fragments_by_id(fragment_ids, fragments_db_filename_or_url, prefix=''):
 
     Returns:
         pandas.DataFrame: Data frame with fragment information
+
+    Raises:
+        IncompleteFragments: When one or more of the identifiers could not be found.
     """
     if fragments_db_filename_or_url.startswith('http'):
         client = WebserviceClient(fragments_db_filename_or_url)
-        fragments = client.fragments_by_id(fragment_ids)
+        try:
+            fragments = client.fragments_by_id(fragment_ids)
+        except IncompleteFragments as e:
+            df = pd.DataFrame(e.fragments)
+            df.rename(columns=lambda x: prefix + x, inplace=True)
+            raise IncompleteFragments(e.absent_identifiers, df)
     else:
         fragmentsdb = FragmentsDb(fragments_db_filename_or_url)
-        fragments = [fragmentsdb[frag_id] for frag_id in fragment_ids]
+        fragments = []
+        absent_identifiers = []
+        for frag_id in fragment_ids:
+            try:
+                fragments.append(fragmentsdb[frag_id])
+            except KeyError:
+                absent_identifiers.append(frag_id)
+        if absent_identifiers:
+            df = pd.DataFrame(fragments)
+            df.rename(columns=lambda x: prefix + x, inplace=True)
+            raise IncompleteFragments(absent_identifiers, df)
 
     df = pd.DataFrame(fragments)
     df.rename(columns=lambda x: prefix + x, inplace=True)